requirements.txt
colorlog==2.10.0
control==0.7.0
matplotlib==2.0.0
numpy==1.13.3
PyYAML==3.12
slycot==0.2.0
